docker-steps/dataset/requirements.txt
bio==1.6.0
biopython==1.81
biothings-client==0.3.0
boto3==1.28.75
botocore==1.31.75
certifi==2023.7.22
charset-normalizer==3.3.2
checksumdir==1.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.43.1
fsspec==2023.10.0
gprofiler-official==1.0.0
huggingface-hub==0.17.3
idna==3.4
importlib-resources==6.1.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
mygene==3.2.2
networkx==3.1
numpy==1.24.4
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
platformdirs==3.11.0
pooch==1.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
s3transfer==0.7.0
sacremoses==0.1.1
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
zipp==3.17.0
